Moved page_scaffold components into page_scaffold folder
