+++ client/flutter/lib/pages/question_index.dart
@@ -1,7 +1,7 @@
 
 import 'dart:math';
 import 'package:WHOFlutter/api/question_data.dart';
-import 'package:WHOFlutter/components/page_scaffold.dart';
+import 'package:WHOFlutter/components/page_scaffold/page_scaffold.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_html/flutter_html.dart';
