+++ client/flutter/lib/pages/who_myth_busters.dart
@@ -1,6 +1,6 @@
 import 'package:WHOFlutter/components/emojiHeader.dart';
 import 'package:WHOFlutter/components/listItem.dart';
-import 'package:WHOFlutter/components/page_scaffold.dart';
+import 'package:WHOFlutter/components/page_scaffold/page_scaffold.dart';
 import 'package:WHOFlutter/generated/l10n.dart';
 import 'package:flutter/material.dart';
